revise: reorder connection_profiles.ui to reduce git noise + other mods
Returned original order to QGridLayout items in:
src/ui/connection_profiles.ui to minimise the git diff/patch noise
introduced by the prior commits into the PR that Vadim created that this is
to be inserted into.

Older format of file also allows all four margin aspects to be specified as
a single entity which is respected by current uic compiler even though the
later Qt Designer plugin insists on redoing them as four separate elements
even if they are all the same (0) as it happens.

Also simplified the HTML formatting used for the "Welcome To Mudlet" text
displayed upon first use.  The text content should be the same but the
markup is simpler so will be easier on any translators who have to deal
with the raw HTML if they look at the content in Qt Linguist in the future.

Revised dlgConnectionProfiles::writeProfileData(...) so it does not produce
QWarning() messages when trying to write profile data for non-existent
profiles - as happens for predefined MUDs that have never actually been
played every time they are clicked on to view in the dialogue!  It does now
return a QPair that indicates success or failure and a QString being the
QFile::errorString() from the failed command.  These are not currently used
but might be in the future.

Abstracted accessing the "Blurb" or description for the predefined MUDs
which is keyed to the host server url (and for the port for the 3k.org case
where two MUDs use the same server).  This is a temporary step so that the
following enhancement can be made but eventually I intend to abstract all
the server settings to a "profile.ini" file accessed through QSettings in
the profiles' individual directories so that new predefined ones can be
installed separately and existing ones can be edited if required.

Added the "blurb" for the predefined MUDs to a tool-tip so that the user
can find out more about those MUDs that we have something for even when
they have not clicked on them in normal usage OR when the Description
widget is hidden on "first" usage after install when the behaviour has
recently been modifed.

Re-positioned the Warning dialog that pops-up when an in-use profile is
selected to be at the bottom of the MUD profile icons QListWidget - when it
was positioned at the top it would cause them all to be repositioned which
is disturbing given that clicking on one of them is what makes the dialogue
pop-up in the first place...!

Tweaked the overall sizes used for both the "first use after install" and
the "normal" case so that there is space to show all the predefined MUD
icons AND the warning dialog should the "default host" be selected.

Renamed: (QListWidget *) dlgConnectionProfiles::profiles_tree_widget
to:      (QListWidget *) dlgConnectionProfiles:listWidget_profiles
as it was confusingly using the wrong widget type in the name.

Reformatted the source code for the method I have directly created/edited
with this commit to our current clang-format settings - apart from those
only edited with the rename of the member to "listWidget_profiles".

Added tool-tips with some salient details to the entry fields on the right
side of the dialog.

Reverted previous removal of initial usage of status bar.  It is needed to
prime the status bar for display of "statusTips" if they are needed later,
so cannot be skipped even if it is set to be disabled in the global
settings initially.  Have revised the wording used to reflect the automagic
display of the connection preferences dialog on start-up.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>