BugFix: make changed showTabs/Spaces & showLineFeed/Paragraphs immediate
Immediate that is in the preview "Editor" widget in the Profile Preference
(settings/option) dialog - changes to the main Editor (for all and each
profile) are not made until the dialog is closed with a save but as the
Editor built into one tab is supposed to be a preview it needs to change
as the options are.

Note that currently showLineFeeds/Paragraphs is not like the original
Qt Text Editor widget - instead it just puts a graphical line between each
line of text.  I believe it is possible to revise this upstream in the
Edbee-lib repository to be more like the LineEnd/Paragraph markings of the
original Qt widget but that will take a bit of time to do that I do not
have right now.

Also note that changing the show tabs / spaces option did not take effect
immediately in the Edbee editor widget but I have put forward a bugfix for
that which has already been accepted into upstream as:
PR https://github.com/edbee/edbee-lib/pull/33 as a fix for:
issue https://github.com/edbee/edbee-lib/issues/32

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>