Revise: ensure all aspects of edbee settings persist on item change
During testing I found that previous changes to the edbee widget set in the
profile preferences did not stay when a new Mudlet item was selected in the
editor and that a theme setting made in that dialog was routed to ALL
profiles instead of the active one.

The first of these issues I solved by ensuring all of the Font, the Theme
and the two global settings for show tabs/spaces and (the currently
misused) linefeed/paragraph marks are all set when
dlgTriggerEditor::clearDocument(...) was called.

The second was fixed by removing the inappropriate
(void) mudlet::setEditorTheme(const QString&) method and the SIGNAL
signal_editorThemeChanged(QString); the latter was wired up to a lambda
method in the dlgTriggerEditor constructor, what it did was needed but only
as a method called by (void) dlgProfilePreferences::slot_save_and_exit()
which was named
(void) dlgTriggerEditor::setThemeAndOtherSettings(const QString &).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>