BugFix: tweak some issues when not wanting to include updater code.
Correct compilation failure when INCLUDE_UPDATER is NOT defined in
(void) dlgProfilePreferences::slot_save_and_exit()

Correct previous qmake / cMake project files which no longer behave as
required as there is now more than one git sub-module.  Also allows command
line building to suppress the effects of the INCLUDE_UPDATER compiler
#define value by specifying an environment variable NO_INCLUDE_UPDATER to
a non-empty string (or a number) before the relevant (makefile) building
command is run, typically this would be from the parallel build directory:
* For qmake (gcc):
  NO_INCLUDE_UPDATER="string" sh -c 'qmake -spec linux-g++ ../src/'
  or qmake (clang):
  NO_INCLUDE_UPDATER="string" sh -c 'qmake -spec linux-clang++ ../src/'
* For cmake (gcc):
  NO_INCLUDE_UPDATER="string" CC=/usr/bin/gcc CXX=/usr/bin/g++ cmake ..
  or cmake (clang);
  NO_INCLUDE_UPDATER="string" CC=/usr/bin/clang CXX=/usr/bin/clang++ cmake ..

CMake will read the environment from the command line but qmake (possibly
because on my system it could be invoked via a symbolic link or a shell
wrapper) does not seem to notice environment settings preceding it...?!

Change (bool) mudlet::self()->onDevelopmentVersion() function into a
(const static bool) mudlet::scmIsDevelopmentVersion value which is faster
and more compact IMVHO.

Add .gitmodules to files visible in Qt Creator as it is not something that
ought to be overlooked.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>