Revise: make Discord RPC library a compilation option on supported platforms
In order to isolate the code from Platforms that do not support Discord
(currently FreeBSD and 32-Bit Linux) I needed to be able to exclude some
parts of the code - and it then made sense to use the same system as has
been adopted for the Ubuntu (and other) fonts and the Updater.

This will prevent inclusion of the Discord code if the Environmental
variable WITH_DISCORD is set to a case insensitive "no" during compilation.

Attempts to make CMake work in this manner was initially tried on Windows but it
seems that the CMake project file is defective in other areas.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>