Revise: further developments
Remove Vadim's original Lua and Discord support functions for:
* `setDiscordGamegame`
* `setDiscordArea`
* `setDiscordCharacter`
* `setDiscordCharacterIcon`

Adds methods/lua functions to set:
* Elapsed/Remaining Timers
* PartySize/Maximum details (note that a party size of zero for a non-zero
  max size is not supported upstream).
* check stored details of a Discord User Name and Number against those
  returned from the Discord RPC library - and if given and not matching
  completely prevents the profile from using Discord. This is a safety
  feature to prevent a Discord user with more than one account accidentally
  trying to connect when they are not logged into Discord with the matching
  user name!

Adds a per item control for (initially GMCP only) access to each of the
Rich Presence details - so that any (or all) of them can be disabled from
outside setting - or for the Discord User Name, for reading. The control
for the Rich Presence Id only controls the writable aspect of the override
from the Mudlet default.  To enable the Server to know how to use the
Rich Presence it HAS to know what PresenceId is in use and it is now sent
that as part of the initial GMCP "External.Discord.Hello".

Also:
* Eliminate some unnecessary `QGridLayout`s of single column or rows of
widgets from the Profile Preferences dialogue/form and switch them to
`QVBoxLayout`s as those can do the same job without having to worry about
maintaining the row/column order to keep the Git noise on editing the
design - the widgets are implicitly listed in the XML file in the order in
which they are placed...!
* Reorder the remaining QGridLayouts in the Profile Preferences dialogue
into a "proper" order again.
* Adds and uses a static QString mudlet::htmlWrap(QString) method to
  eliminate repeated use of a commonly use QStringLiteral(...) that puts
  the HTML necessary around tool-tips to convert them from "plain-text" to
  "rich-text" - which is handled better for tooltips of more than a few
  words.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>