Revise: insert QtKeyChain licence boilerplate
I have use the 2-Clause BSD (a.k.a the FreeBSD) licence that is in the
upstream's "COPYING" file - this is at varience with the two README files
which quote the licence as being the Modified BSD which is the 3-Clause
one. I have an open issue on the upstream repository seeking clarification;
should things not be as they seem it should be a relatively painless
operation to select the other boillerplate for this item.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>